requirements.txt
digitalai-release-sdk==25.3.0
